Add initial project files via upload
# ViridianKrypton

ViridianKrypton is a work-in-progress class project for:
 Microsoft Software Systems Academy (MSSA) -
  Cloud Application Developer (CAD) course
   as administered by Embry-Riddle Aeronautical University

## Description

BLUF: The goal of this project is to develop a centralized repository of 3D scans of people in order to manage this highly-sensitive biometric resource while still providing secure, managed access for authorized users.

Situation: 
At present, full-body 3D scanning is a relatively niche experience limited primarily to medical diagnostics such as MRIs, high-end fitness clubs, and airport security scans. 3D scanning hardware is constantly improving and hardware prices are beginning to drop into the hobbyist price range. Soon, this capability will be ubiquitous just as it has become commonplace to log into a computer or phone through facial recognition. 

I believe the "killer app" for 3D body scans will be in online clothing sales by allowing retailers to return customized search results and recommendations for clothing guaranteed to fit and flatter the customer. No more sizing charts, label-specific sizing, or clothing that looks great on a fashion model but won't fit 90% of the population correctly. Other applications spring to mind such as personalized gaming avatars (both video and board game), public health policy recommendations, and medical research.

In the era of DeepFakes, the thought of widespread, unregulated availability of this kind of data is unnerving. A solution is required to collect and secure this data, while still allowing individuals to choose to release it for authorized purposes (or to delete it altogether). 


## Contributing
Pull requests are welcome. For major changes, please open an issue first to discuss what you would like to change.

Please make sure to update tests as appropriate.

## Resources
Sample 3D Scans  were produced using [DesktopHero 2.0](https://desktophero3d.com/) and edited in Microsoft 3D Viewer; no humans were injured nor imaged in the making of this project.

## License
Copyright: Kurt Woodward (2021)
All rights reserved.